Classify: Correct the margins/padding.
Now the Classify activity has no padding or margins.
The Classify Fragment's ScrollView has no margin or padding at
its sides, so the scrollbar shows up at the far edge, but it has
a margin at the top to separate it from the ActionBar.
This seems to achieve this layout:
https://developer.android.com/design/style/metrics-grids.html
The ScrollView's LinearLayout has a margin at the sides to separate
the content from the sides (and from the scrollbar).