Classify: Use clipToPadding to scroll into the top gap.
Change the activity's top margin to top padding in the fragment's
ScrollView, so we can use clipToPadding="false" to make it scroll
its content into the gap between the content and the ActionBar.
This only works if the padding is on the ScrollView (or ListView
or GridView, I think) itself.
See https://plus.google.com/+AndroidDevelopers/posts/LpAA7q4jw9M