Tablet layout: use layout-w850dp instead of layout-sw-600dp.
sw means the smallest of width or height.
w means the width in the current orientation.
This then means Use this layout if the current width is more than
850dp.
This makes it use the standard vertical layout on my Nexus 10 when
in portrait mode, so we can see a large picture because the picture
has enough width to expand.
The 850 is odd, but seems to be what works.