QuestionFragment: Make the TableLayout columns have equal widths.
update(): Use width=0 and weight=1 for the TableRow's children, and use
width=MATCH_PARENT for the TableRow, to make all columns have equal size.
Note that if width=WRAP_CONTENT then the TableRow's children will be
completely invisible.
And add extra placeholder FrameLayouts to fill the remaining cells in the last
column if there is more than one column.