ItemsContentProvider.requestMoreItemsSync(): Do not request too many.
In order to avoid too much network and disk IO all at the same time
as we download the images. This can make the UI non responsive.
It doesn't seem to help to do it in a Thread with Thread.MIN_PRIORITY
either - too much IO will just slow the UI down.