ItemsContentProvider: Don't do regular tasks (I/O) until UI seems to be idle.
queueRegularTasks(): Instead of scheduling the work for x seconds after it
was requested, keep waiting until there have been x seconds without a request.
None of the work is urgent apart from getting at least one item to classify,
and we already get one synchronously if that is really necessary.
This should reduce the chance of the I/O happening while the user is doing
classifications, reducing the chance that the user experiences the UI being
momentarily non responsive.

Also reduce the delay from 20 seconds back to 10 seconds.