Change "Loading…" to "Downloading images…".
It's not really what it's doing. In particular, on subsequent starts
it's really just parsing its local XML and checking the timestamps of the
icons and example images that it downloaded the first time.
But Loading… seems generic enough, and it takes long enough, that people
don't have confidence that it's ever going to finish.
Even "Downloading images…" is too much technical detail to ever put in
a UI, I think, so maybe we'll just have to put the actual icons and
example files in the executable, with the downside that the
package would have to be rebuilt and re-uploaded to the Play Store
every time they are changed on the website. We already do that
for the decision tree (and its translations), but that seems less
likely to change.