In portrait mode, try to reduce the changes of image height.
By remembering how big the question/answer part has ever been,
and making it always be that big. This means that the image resizes
during the first classification, but not afterwards.
It will resize again if you don't use the "Is there anything Odd?"
question until later.
This is far from ideal.