QuestionFragment: Add an invisible second row of answers when necessary.
This makes each question's height more similar, though the actual height still
depends on the actual text (which can vary in unexpected ways when using
translations) and the actual icon (which is unlikely to be unexpected).
I don't like hard-coding the number 2, but the decision tree is unlikely to
change much. This would only be a problem when adapting this code to
a different decision tree.