IconsCache: Bundle icons from assets and use them if they are new enough.
However, the next commit will make us always use them from assets
and never try to download them, because the CSS file (which describes
which parts of the icons file to use for each icon) changes
for each release, and I don't want to bother with coping wth that,
at least for now.

Note that these icons are extracted from these files:
https://github.com/zooniverse/Galaxy-Zoo/blob/master/public/images/workflow.png
https://github.com/zooniverse/Galaxy-Zoo/blob/master/public/images/examples.jpg
which are under a compatible license (Apache License Version 2.0).