ClassifyActivity: listen to preferences changes here instead of in SyncAdapater.
Because it doesn't work in SyncAdapter when the SyncAdapter is in its own
process (when AndroidManifest.xml's <service> for it has
android:process=":sync" ). This is semi-documented: The SharedPreferences
docuementation says
"Note: currently this class does not support use across multiple processes. This will be added later."
However, do _get_ the correct preferences values in the SyncAdapter even
if we cannot listen for changes to them.