Toolbar: Use a derived style to get a dark menu.
Instead of just using Theme.AppCompat.Light to get the dark title text
and overflow icon in the toolbar, use a special theme for the toolbar
that derives from Theme.AppCompat (with the dark menu) and change
textColorPrimary and textColorSecondary to get the dark title text
and overflow icon.

However, the menu colors are not quite right, because those two
values affect it too. But setting popupTheme on the <Toolbar>
lets us change the menu's theme back again, leaving our customization
only for the toolbar itself.
https://chris.banes.me/2014/10/17/appcompat-v21/