Classify: Toggle buttons: Don't use a custom style.
Because, even when setting textColor, textSize and textStyle to @null,
we still don't get quite the same appearance as with textAppearance on
the Button itself. Incidentally, for some reason using
android:textStyle="@null"
fails style validation though it works for textColor and textSize.
I assume that the appearance is more correct when using the
textAppearance directly on the Button in the layout.