Toolbar: Replace the title text with the (word mark) icon.
It is now even more obvious that the title (now our icon containing text)
is black while the action items are grey.
However, this seems to be the default in the Material (and Holo)
dark theme. Google's Photo app has the same colors, for instance,
though I have not found another suitable example.