AndroidManifest: READ_EXTERNAL_STORAGE: Use maxSdkVersion="18"
Our use of WRITE_EXTERNAL_STORAGE has this already, because
use of our app-private getExternalCacheDir() doesn't need it in
newer API versions. But WRITE_EXTERNAL_STORAGE implies
READ_EXTERNAL_STORAGE, but the maxSdkVersion doesn't seem to be
carried over to the implied permission. I noticed this in
the Google Play console.