Change ic_action_help to black, instead of gray.
This is not the default, but it just looks bad with (default) black
title text, and plenty of official google apps seems to change the
icons to be fully white or fully black.
Generated with:
http://romannurik.github.io/AndroidAssetStudio/icons-actionbar.html#source.type=clipart&source.space.trim=0&source.space.pad=0&source.clipart=res%2Fclipart%2Ficons%2Faction_help.svg&name=ic_action_help&theme=custom&color=000%2C100

I wish this the Material theme had better defaults and I wish this
was mentioned in the Material design guidelines.