ClassifyActivity: Fix a crash in logged-in check after 3 classifications.
Call our LoginUtils.getLoggedIn() from an ASyncTask.
A user reported a crash via Google Play with this exception:
  java.lang.IllegalStateException: calling this from your main thread can lead to deadlock
Strangely, that looks like a StrictMode exception, though StrictMode
is not enabled in the app, and I can't reproduce it when I enable StrictMode,
but the code really should be in an AsyncTask anyway.