ClassifyActivity: Avoid adding the ClassifyFragment twice.
Because I think this is sometimes happening on Android 5.0 (Lollipop).
For instance, sometimes all the action icons (and overflow menu items) are doubled,
and the question text has other questoin text underneath it.
This might be when the app has been in the background for some time
and is then reopened.
If I could find how to reproduce it reliably then I could investigate properly.