ExampleViewerActivity: Make the Toolbar's Up Button go back properly.
Override onOptionsItemSelected() so we can modify the Intent
used with navigateUpTo(), to specify FLAG_ACTVITY_CLEAR, so we
resume the parent activity instead of recreating it, just as if the
user had used the back button - which the user generally cannot
distinguish from the up button anyway.
We could instead have marked the parent activity as singleTop
in the manifest, but it seems nice to keep this parent activity
more flexible.

This should fix this bug:
https://github.com/murraycu/android-galaxyzoo/issues/14