QuestionFragment.ClassificationInProgress: Do some defensive copying.
Return copies of lists and copy lists in constructors.
This is to avoid concurrent changes when, for instance, the answers
List changes while we are iterating over it. It's not clear when that
might happen, but we call getAnswers() from AsyncTask(), so it seems
possible. There is at least one report of a ConcurrentModificationException
so this might help:
https://github.com/murraycu/android-galaxyzoo/issues/15