SubjectFragment.onLoadFinished(): Call destroyLoader() to avoid extra calls.
As we do in our other onLoadFinished() methods, to avoid an apparent Android
bug that causes multiple extra calls. This was causing showImage() to be
called with some older out-of-date item IDs, confusing our checks, as
well as being inefficient.

This seems to avoid the problem where the user is left with just a blank
image (and the question) after the UI has had to abandon some items
due to their images being deleted from the cache. Now it doesn't stop until
it has a new image or until it has complained about a lack of a suitable network
connection.

Hopefully this completely solves this bug:
https://github.com/murraycu/android-galaxyzoo/issues/12