LoginUtils.removeAccount(): Avoid calling new API on older Android versions.
Use the older (deprecated in API 22) AccountManger.removeAccount() method
on older Android versions, to avoid the NoSuchMethodError exception
(and crash) when doing this on Android versions older than 5.1
(almost all installations at the moment).
We don't have this problem with most other API because we use AppCompat
API for so much, but there is no AppCompat for AccountManager.