ExampleViewer: Take and try an alternative URL.
* DecisionTree: Add getExampleIconNameWithCommonMistake(). Often the - after
the survey group name, before the question number, is a _ in the example
file name. That breaks the web UI's ability to show full example images too.
But this lets us try those URLs instead of waiting for the filenames
to be corrected on the server.
* ExampleViewerActivity, ExampleViewerFragment: Take an alternative URI and
try it if the first one failed.
* QuestionHelpFragment: Pass the alternative URL.

See https://github.com/zooniverse/Galaxy-Zoo/pull/198