Revert "Correct sharing of image via share menu item."
This reverts commit 863c54e20969b1f542216ce998c4946cc4c1c234.

Because SubjectFragment.updateShareActionIntent() is always called from
onCreateOptionsMenu(), and updateShareActionIntent() always executes the
GetImageBitmapAsyncTask, getting the image just in case the user wants
to share it.

That's unnecessary work in most cases and it causes us to ask the user
for photos/files permission. We really should delay that permission
request (and the actual downloading) until the user actually tries to
share something.